adds support for execution as an NPM package
