+++ package.json
@@ -4,12 +4,13 @@
     "uglify-js": "^3.17.4"
   },
   "name": "spiraappvalidator",
-  "version": "1.0.0",
-  "main": "bundler.js",
+  "version": "2.0.0",
+  "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "node bundler.js bundle $INPUT $OUTPUT"
+    "build": "node index.js $INPUT $OUTPUT"
   },
+  "bin": "./index.js",
   "author": "",
   "license": "ISC",
   "description": ""
